security: move triggerAIDiffAnalysis out of "use server" (IDOR)
triggerAIDiffAnalysis was an unguarded server action: any authenticated user
could invoke it with any buildId, spending the owning team's AI budget and
writing analysis results onto their diffs. It now lives in
@/lib/ai/trigger-diff-analysis as a plain function imported by trusted callers
(build finalization, guarded actions).
